requirements.txt
aws==0.2.5
bcrypt==3.2.0
blueprint==3.4.2
boto==2.49.0
boto3==1.21.0
botocore==1.24.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
fabric==2.6.0
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
invoke==1.6.0
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
paramiko==2.9.2
pathlib2==2.3.7.post1
prettytable==3.1.1
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
